// types/database.d.ts
export type position = "bench" | "gk" | "def" | "mid" | "att";
// ...
export interface leagueSettings {
  leagueName: string;
  leagueID: number;
  startMoney: number;
  transfers: number;
  duplicatePlayers: number;
  starredPercentage: number;
  league: string;
  archived: number;
  matchdayTransfers: boolean;
  fantasyEnabled: boolean;
  predictionsEnabled: boolean;
  predictWinner: number;
  predictDifference: number;
  predictExact: number;
  top11: boolean;
  active: boolean;
  inActiveDays: number;
}
// ...
export interface transfers {
  leagueID: number;
  seller: number;
  buyer: number;
  playeruid: string;
  value: number;
  position: position;
  starred: boolean;
  max: number;
}